Few new attributes on runtimes and files
Allow an assembly to specify and outputDirectory, to handle the case where there may be multiple provisio executions and we don't want everything landing in the same directory.

Allow a file to specify an "as" attribute like with artifacts so they can be written out with a different name if desired.